web: refactor 7/7 main js
